Split tests into feature for readability
